Refactored names of testable widgets
