+++ app_feup/test/unit/view/Pages/ExamsPageView_test.dart
@@ -3,7 +3,7 @@ import 'package:app_feup/view/Pages/ExamsPageView.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
 
-import '../../TestableWidget.dart';
+import '../../../TestableWidget.dart';
 
 void main() {
   group('ExamsPage', () {
