+++ app_feup/test/unit/view/Widgets/ScheduleRow_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/material.dart';
 
 import 'package:app_feup/view/Widgets/ScheduleRow.dart';
-import '../../TestableWidget.dart';
+import '../../../TestableWidget.dart';
 
 void main() {
   group('ScheduleRow', () {
